requirements.txt
flask
flask-restful
flask-jsonpify
pythainlp==1.6.0.4
gunicorn
numpy
gensim
cython
fasttext
nltk
pyemd
psycopg2
flask-httpauth
passlib
itsdangerous
yoyo-migrations
pylint
autopep8